+++ src/ColorSpecialHandler.cpp
@@ -49,7 +49,7 @@ static void hsb_to_rgb (const vector<float> &hsb, vector<float> &rgb) {
 		rgb[0] = rgb[1] = rgb[2] = hsb[2];
 	else {
 		float h = hsb[0]-floor(hsb[0]);
-		int i = 6*h;
+		int i = (int) (6*h);
 		float f =  6*h-i;
 		float p = hsb[2]*(1-hsb[1]);
 		float q = hsb[2]*(1-hsb[1]*f);
