+++ src/PsSpecialHandler.h
@@ -82,8 +82,8 @@ class PsSpecialHandler : public SpecialHandler, protected PSActions
 		void setdash (std::vector<double> &p);
 		void setgray (std::vector<double> &p);
 		void sethsbcolor (std::vector<double> &hsb);
-		void setlinecap (std::vector<double> &p)     {_linecap = p[0];}
-		void setlinejoin (std::vector<double> &p)    {_linejoin = p[0];}
+		void setlinecap (std::vector<double> &p)     {_linecap = (unsigned int) p[0];}
+		void setlinejoin (std::vector<double> &p)    {_linejoin = (unsigned int) p[0];}
 		void setlinewidth (std::vector<double> &p)   {_linewidth = p[0] ? p[0]*1.00375 : 0.5;}
 		void setmatrix (std::vector<double> &p);
 		void setmiterlimit (std::vector<double> &p)  {_miterlimit = p[0]*1.00375;}
